requirements.txt
# requirements.txt
moviepy==1.0.3
google-generativeai==0.1.0
selenium==4.8.0
python-dotenv==1.0.0
gTTS==2.2.3
PyYAML==6.0
requests==2.28.1
google-auth-oauthlib==0.7.0
google-auth-httplib2==0.1.0
google-api-python-client==2.70.0
